windows installer and logging
