+++ .github/workflows/release.yml
@@ -164,20 +164,52 @@ jobs:
           cd dui
           deno task build
 
+      # Import and setup Apple certificates for macOS signing
+#     - name: Import Apple Developer Certificate
+#       if: matrix.platform == 'macos'
+#       env:
+#         APPLE_CERTIFICATE: ${{ secrets.APPLE_CERTIFICATE }}
+#         APPLE_CERTIFICATE_PASSWORD: ${{ secrets.APPLE_CERTIFICATE_PASSWORD }}
+#         KEYCHAIN_PASSWORD: ${{ secrets.KEYCHAIN_PASSWORD }}
+#       run: |
+#         echo $APPLE_CERTIFICATE | base64 --decode > certificate.p12
+#         security create-keychain -p "$KEYCHAIN_PASSWORD" build.keychain
+#         security default-keychain -s build.keychain
+#         security unlock-keychain -p "$KEYCHAIN_PASSWORD" build.keychain
+#         security import certificate.p12 -k build.keychain -P "$APPLE_CERTIFICATE_PASSWORD" -T /usr/bin/codesign
+#         security set-key-partition-list -S apple-tool:,apple:,codesign: -s -k "$KEYCHAIN_PASSWORD" build.keychain
+#         security find-identity -v -p codesigning build.keychain
+#
+#     - name: Verify Certificate
+#       if: matrix.platform == 'macos'
+#       run: |
+#         CERT_INFO=$(security find-identity -v -p codesigning build.keychain | grep "Apple Development")
+#         CERT_ID=$(echo "$CERT_INFO" | awk -F'"' '{print $2}')
+#         echo "CERT_ID=$CERT_ID" >> $GITHUB_ENV
+#         echo "Certificate imported."
+
       - name: Build DUI (Tauri)
         uses: tauri-apps/tauri-action@v0
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           MACOSX_DEPLOYMENT_TARGET: "10.13"
+          # macOS signing environment variables
+          APPLE_CERTIFICATE: ${{ secrets.APPLE_CERTIFICATE }}
+          APPLE_CERTIFICATE_PASSWORD: ${{ secrets.APPLE_CERTIFICATE_PASSWORD }}
+          APPLE_ID: ${{ secrets.APPLE_ID }}
+          APPLE_ID_PASSWORD: ${{ secrets.APPLE_ID_PASSWORD }}
+          APPLE_SIGNING_IDENTITY: ${{ env.CERT_ID }}
         with:
           projectPath: './dui'
           args: ${{ matrix.platform == 'macos' && '--target universal-apple-darwin' || format('--target {0}', matrix.target) }}
+          #args: ${{ matrix.platform == 'macos' && '--target universal-apple-darwin' || format('--target {0}', matrix.target) }} --debug
 
       # Process and upload build artifacts
       - name: Process build artifacts
         shell: bash
         run: |
           BUNDLE_DIR="dui/src-tauri/target/${{ matrix.target }}/release/bundle"
+          #BUNDLE_DIR="dui/src-tauri/target/${{ matrix.target }}/debug/bundle"
           echo "Build output directory contents:"
           ls -R "$BUNDLE_DIR" || echo "No bundle directory found"
