+++ api/src/main.ts
@@ -136,7 +136,8 @@ const cleanup = async (code: number = 0) => {
 		Deno.exit(1);
 	}
 };
-const signals: Deno.Signal[] = ['SIGINT', 'SIGTERM'];
+// Windows only supports SIGINT and SIGBREAK, while Unix-like systems support SIGINT and SIGTERM
+const signals: Deno.Signal[] = Deno.build.os === 'windows' ? ['SIGINT', 'SIGBREAK'] : ['SIGINT', 'SIGTERM'];
 for (const signal of signals) {
 	Deno.addSignalListener(signal, cleanup);
 }
