+++ dui/src-tauri/Cargo.toml
@@ -34,7 +34,7 @@ once_cell = "1.19"
 log = "0.4"
 env_logger = "0.10"
 log4rs = { version = "1.3", features = ["all_components", "yaml_format"] }
-chrono = "0.4"
+chrono = { version = "0.4", features = ["serde"] }
 http = "0.2"
 hyper = { version = "0.14", features = ["full", "http1", "http2", "client"] }
 hyper-tls = "0.5"
@@ -57,5 +57,11 @@ windows-sys = { version = "0.48", features = [
     "Win32_System_Threading",
     "Win32_System_Diagnostics_Debug",
     "Win32_Security",
-    "Win32_System_ProcessStatus"
+    "Win32_System_ProcessStatus",
+    "Win32_System_Console",
+    "Win32_UI_WindowsAndMessaging",
+    "Win32_System_LibraryLoader",
+    "Win32_System_EventLog",
+    "Win32_System_Registry",
+    "Win32_System_Com"
 ] }
