+++ dui/src-tauri/tauri.conf.json
@@ -56,8 +56,8 @@
         "type": "embedBootstrapper"
       },
       "wix": {
-        "language": "en-US",
-        "enableElevatedUpdateTask": true
+        "upgradeCode": "C8A8C5A6-5BF5-40D6-9E4F-632D98C26F2D",
+        "fragmentPaths": ["windows/fragments/registry.wxs"]
       },
       "nsis": {
         "installMode": "both",
